examples/rfnry-chat/stock-assistant/server-python/pyproject.toml
[project]
name = "rfnry-chat-example-stock-assistant-server"
version = "0.0.1"
description = "Stock assistant example — standalone chat server"
requires-python = ">=3.11"
dependencies = [
    "rfnry-chat-protocol",
    "rfnry-chat-server",
    "fastapi>=0.136.0",
    "uvicorn[standard]>=0.44.0",
    "python-dotenv>=1.2.1",
]

[tool.uv.sources]
rfnry-chat-protocol = { path = "../../../../../types/packages/chat-python", editable = true }
rfnry-chat-server = { path = "../../../../../chat/packages/server-python", editable = true }

[project.optional-dependencies]
dev = [
    "mypy>=1.20.1",
    "poethepoet>=0.44.0",
    "ruff>=0.15.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.format]
quote-style = "double"

[[tool.mypy.overrides]]
module = ["socketio", "socketio.*"]
ignore_missing_imports = true

[tool.poe.tasks]
dev = "uvicorn src.main:asgi --reload --host 0.0.0.0 --port 8000"
start = "uvicorn src.main:asgi --host 0.0.0.0 --port 8000"
format = "ruff format ."
"format:imports" = "ruff check --select I --fix ."
check = "ruff check ."
"check:fix" = "ruff check --fix ."
typecheck = "mypy src"
